mon: do not pollute directory with cvs files from crushtool
The --output-csv option to crushtool will create files in the current
directory of the monitor. The only reason for using it is because
crushtool requires at least one option for display. Relax this
constraint in crushtool and remove the option from the call made by the
monitor to validate a new crushmap.

Signed-off-by: Loic Dachary <ldachary@redhat.com>