tests: keep intermediate docker build layers
Because the docker devmapper backend has an intermittent bug

   https://github.com/docker/docker/issues/4036

the build of the image will sometime fail. It will eventually succeed if
the command is repeated enough times. Not removing the intermediate
layers helps speed up the operation: when repeating the command, only
the layer that failed because of the above bug is rebuilt.

Signed-off-by: Loic Dachary <ldachary@redhat.com>