tests: increase the delay waiting for a cluster event
In some cases the machine running tests may be slow enough that it takes
more than a minute for an OSD to come up. It only happens rarely and
changing the wait period from 60 seconds to 120 seconds. Is so slow
that it takes more than 2 minutes to bring an OSD up, chances are a lot
more will go wrong anyway.

Signed-off-by: Loic Dachary <ldachary@redhat.com>