PG: prefer newer last_epoch_started
We want to make sure that last_epoch_started even in the case where
we go peered instead of active is non-decreasing for the participants
in an interval.  This may have caused 11110:
1) epoch 10, [1, 2] go active les = h.les = 10, lu = 10.13
2) epoch 11, [1] goes active les = h.les = 10, lu = 10.13 (no updates)
3) epoch 12, [1, 2] starts to go active with 2's info, 1's
last_epoch_started is overwritten to 10, but h.last_epoch_started
remains at 12.
4) PG peers again before going active and since no osd has les >= h.les,
pg remains incomplete.

This patch also adds asserts verifying that last_epoch_started does
not decrease on an osd.

Fixes: 11110
Signed-off-by: Samuel Just <sjust@redhat.com>