whitelist shard missing errors
If a shard is missing during a radosbench it can safely be
ignored. It will either be recovered and radosbench will complete
successfully. Or there are too many shard missing to recover the object
and radosbench will fail. Either way we can rely on radosbench to
accurately report success or failure.

Signed-off-by: Loic Dachary <loic@dachary.org>