ceph.spec.in: rm EOL Fedoras; add OBS RHEL5 instead
Fedora 12 has been EOL for a long time. Remove the reference in the
RPM .spec file.

Since RHEL 5 support for Ceph is a work in progress, we won't remove
this entire python_sitelib / python_sitearch conditional for now, since
those are still needed on RHEL 5.

Add the rhel_version macro to make the conditional compatible with
SUSE's OBS.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>