common, global: use lttng ust functions for handling fork-like calls
This allows lttng to work across fork(2) and related calls, including
daemon(3). Without these calls, LD_PRELOAD would need to include
liblttng-ust-fork.so, which wraps fork and daemon.

If these calls or the liblttng-ust-fork wrappers are not used, the
atexit(3) handlers registered by lttng may cause a crash when the
process exits. This does not happen with upstart, since it runs
ceph-osd and other daemons in the foreground.

Backport: hammer
Signed-off-by: Josh Durgin <jdurgin@redhat.com>