os/FileStore: use new lfn_rename for collection_move_rename
The link/unlink sequence is error prone because we cannot do an atomic
replacement of the target.  Instead, create a new lfn_rename operation.
The guards here are unchanged, except we no longer have to worry about the
case where we get EEXIST (rename does a replacement), and we don't have to
clean up the old link.

Signed-off-by: Sage Weil <sage@redhat.com>