Add more CLI optionsto ceph-rest-api
In order to containerized ceph-rest-api we need to expose some options
on the CLI directly. We don't want to rely on reading the ceph.conf. In
this commit, we expose all the configuration options as CLI extra
parameters.
Now you can run the API like this:

ceph-rest-api --addr 192.168.0.24:5001 --log-file /dev/null --base-url
/api/1.0/ --log-level debug

Signed-off-by: Sébastien Han <sebastien.han@enovance.com>