mds: clarify MDBalancer::send_heartbeat
The initial is_degraded() check guarantees that
the 'in' set is equal to the 'up' set.  Later,
this calls get_mds_set and assigns it to a variable
called 'up'.

It's clearer to use get_up_mds_set into the variable
called up (this was confusing when debugging #11218
which was itself a result of is_degraded() ignoring
damaged ranks).

Signed-off-by: John Spray <john.spray@redhat.com>