ReplicatedPG::finish_promote: fix snap promote head snaps
If the snaps vector is: 10=[9,5,2]:[4]+head, the backing pool's snaps
vector is 3=[2]:[]+head, and we request clone 4 from the backing pool,
the backing pool will send us head with an empty results->snaps vector.
Actually, clone 4 should be trimmed, but the cache pool does not know
that.  Thus, we should construct an empty snaps vector for that clone.

PartialFix: #11296
Backport: firefly, hammer
Signed-off-by: Samuel Just <sjust@redhat.com>