ECBackend: For recovery, we allow not enough shard to recovery in the first round of read.
We can run into the situation where the missing shards cannot be
recovered from what we have, when and only in the first round of read,
because we cancel the recovery in this case.

Using this assert, we hope don't meet this situation that in the process
of recovering we lost shard.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>