tests: allow multiple users of docker-tests.sh
The docker image created by docker-tests.sh for a given operating system
is parameterized with the user name. If two users on the same machine
try to use the same image, they will compete and fail with an error
like:

... user get supplementary groups Unable to find user ...

Add the $USER to the image name to reflect the fact that they contain an
account for this user.

Signed-off-by: Loic Dachary <ldachary@redhat.com>