librbd: notify_resize doesn't notify change to all clients
When a resize request is sending to a non lock owner client, it calls
notify_resize to send a resize request to the lock owner. However, the
lock owner doesn't notify all the clients about this change after
completing the request. Should send notify change at the end of the
async resize request.

Signed-off-by: Zhiqiang Wang <zhiqiang.wang@intel.com>