release-notes: backport Hammer release notes
Include the Hammer release notes in Hammer.

Cherry picking the Hammer release notes cannot be done cleanly, they are
copy/pasted instead. This will allow cherry-picking the release notes
for the next point releases. It should be undisturbed by the release
notes for other point releases because they modify parts of the file
that will not generate cherry-pick conflicts.

Signed-off-by: Loic Dachary <ldachary@redhat.com>