osd/ReplicatedPG: promote snapdir when both head and snapdir are missing
When the client op is on snapdir, and both the head and snapdir are
missing on the cache tier, we should promote the snapdir, instead of the
head. This is because if we have snapdir, but no head on the base tier,
promoting head will get nothing. And promoting the snapdir will first
look for the head, and then the snapdir, which would be always correct.

Signed-off-by: Zhiqiang Wang <zhiqiang.wang@intel.com>