mon: Monitor: rework scrub (2)
Instead of hanging while scrubbing all the keys in the store, scrub just
a few keys each time and keep scrubbing until we reach go through all
the keys in the store.

This may miss a few keys if proposals are committed in-between scrub
rounds, but we'll get them eventually (if they are not trimmed first).

Fixes: #11773

Signed-off-by: Joao Eduardo Luis <joao@suse.de>