osd: tolerate when pg_pool_t removed_snaps shrinks
The monitor bug #11493 incorrectly cleared the removed_snaps field for a
pool.  This makes the OSD tolerate this case (regardless of what caused it)
and issue a warning instead of triggering an assertion.

Note that the OSD's notion of 'removed_snaps' does not shrink--we continue
to add values only, not remove them.

Signed-off-by: Sage Weil <sage@redhat.com>