obj_bencher: does not accumulate bandwidth that is zero
We never want to accumulate a bandwidth that is zero. A bandwidth of
zero means that the object could not be written in one second (or that
it's the first time in the loop). Even if writing an object takes a very
long time, the bandwidth cannot be zero, it will just be very small.

Fix suggested by @dachary
Fixes: #7401

Signed-off-by: Dmitry Yatsushkevich <dmitry.yatsushkevich@gmail.com>