erasure-code: lrc size test depends on layer semantic
When the lrc layers are defined, the semantic of the D,c and _
characters are defined, the rest is undefined. The test that verifies
the guard against layers of different size uses the A character which
is undefined. Depending on the implementation, the size test could fail
because the A character is undefined and a guard to forbid undefined
characters is added. Replace A with D to make sure the undefined
character A will not interfere with the test.

This may seem nitpicking but it actually caused problems after a code
refactor that will appear in a few commits from here.

Signed-off-by: Loic Dachary <ldachary@redhat.com>